Update SucrStaking.sol
added emergency withdrawal function back in, can only be used to withdrawal the original amount deposited, minus the amount paid out in interest to the users. This allows the creator to remove funds from the pool if it is closed with funds still available, while protecting the user from the creator stealing their deposits.